requirements.txt
alembic==1.7.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
MouseInfo==0.1.3
mysqlclient==2.0.3
Pillow==8.3.2
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
pytweening==1.0.4
PyYAML==5.4.1
requests==2.26.0
selenium==3.141.0
SQLAlchemy==1.4.23
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.5.0
